requirements.txt
flask==2.3.3
pandas==2.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
openpyxl==3.1.2